ai-service/src/services/vital_signs_service.py
import cv2
import numpy as np
from typing import Dict, Any, Optional
import logging

logger = logging.getLogger(__name__)


class VitalSignsService:
    """
    Service for extracting vital signs from video
    This is a placeholder implementation - in production, this would use
    advanced computer vision and ML models for:
    - Heart rate detection (PPG from facial video)
    - Respiratory rate (chest movement analysis)
    - Oxygen saturation estimation
    - Blood pressure estimation (if applicable)
    - Temperature estimation (thermal imaging or other methods)
    """
    
    def __init__(self):
        self.model_loaded = False
        # In production, load pre-trained models here
        # self.load_models()
    
    async def extract_from_video(self, video_bytes: bytes) -> Dict[str, Any]:
        """
        Extract vital signs from video bytes
        
        Args:
            video_bytes: Video file as bytes
            
        Returns:
            Dictionary containing extracted vital signs
        """
        try:
            # Save video to temporary file
            import tempfile
            import os
            
            with tempfile.NamedTemporaryFile(delete=False, suffix='.mp4') as tmp_file:
                tmp_file.write(video_bytes)
                tmp_path = tmp_file.name
            
            try:
                # Process video
                vital_signs = self._process_video(tmp_path)
                return vital_signs
            finally:
                # Clean up temporary file
                os.unlink(tmp_path)
                
        except Exception as e:
            logger.error(f"Error extracting vital signs: {str(e)}")
            # Return None to allow fallback to manual entry
            return None
    
    def _process_video(self, video_path: str) -> Dict[str, Any]:
        """
        Process video to extract vital signs
        This is a simplified implementation - production would use advanced CV/ML
        """
        cap = cv2.VideoCapture(video_path)
        
        if not cap.isOpened():
            raise ValueError("Could not open video file")
        
        fps = cap.get(cv2.CAP_PROP_FPS)
        frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
        duration = frame_count / fps if fps > 0 else 0
        
        # Placeholder for actual vital signs extraction
        # In production, this would:
        # 1. Detect face/body in frames
        # 2. Extract PPG signal for heart rate
        # 3. Analyze chest movement for respiratory rate
        # 4. Use ML models for other vital signs
        
        # For now, return estimated values based on video analysis
        # These would be replaced with actual ML model predictions
        
        heart_rate = self._estimate_heart_rate(cap, fps)
        respiratory_rate = self._estimate_respiratory_rate(cap, fps)
        
        cap.release()
        
        return {
            "heartRate": heart_rate,
            "respiratoryRate": respiratory_rate,
            "oxygenSaturation": None,  # Would require specialized equipment
            "bloodPressure": None,  # Would require specialized equipment
            "temperature": None,  # Would require thermal imaging
            "consciousness": "alert",  # Would use ML model to detect
            "painLevel": None,  # Would use facial expression analysis
            "source": "video",
            "confidence": 0.7  # Confidence score for the extraction
        }
    
    def _estimate_heart_rate(self, cap: cv2.VideoCapture, fps: float) -> Optional[int]:
        """
        Estimate heart rate from video using PPG (Photoplethysmography)
        This is a placeholder - production would use advanced signal processing
        """
        # Simplified PPG-based heart rate estimation
        # In production: extract color changes from facial region, apply FFT
        
        # For demo purposes, return a reasonable estimate
        # Real implementation would analyze color variations in facial region
        return 72  # Placeholder value
    
    def _estimate_respiratory_rate(self, cap: cv2.VideoCapture, fps: float) -> Optional[int]:
        """
        Estimate respiratory rate from chest/abdominal movement
        This is a placeholder - production would track chest movement
        """
        # Simplified respiratory rate estimation
        # In production: track chest/abdominal movement, count cycles
        
        # For demo purposes, return a reasonable estimate
        return 16  # Placeholder value
    
    def load_models(self):
        """
        Load pre-trained ML models for vital signs extraction
        """
        # In production, load models here
        # Example:
        # self.heart_rate_model = load_model('models/heart_rate_model.h5')
        # self.respiratory_model = load_model('models/respiratory_model.h5')
        self.model_loaded = True

