ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
scipy==1.11.4
scikit-learn==1.3.2
httpx==0.25.2
python-dotenv==1.0.0

